backend/app/agent/llm.py
import os
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv

load_dotenv()

def get_llm():
    return ChatGoogleGenerativeAI(
        model="gemini-2.5-flash",
        temperature=0.2,
        google_api_key="AIzaSyBeCfcjsE0-vEtS23s8UmuWO0FHoxBjwyk"
    )
