backend/requirements.txt
# Core framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Environment variables
python-dotenv==1.0.1

# Authentication & security
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# MongoDB (async)
motor==3.3.2
pymongo==4.6.1

# Vector DB
chromadb==0.4.24

# LLM / AI
google-generativeai==0.3.2

# Utilities
pydantic==2.6.4
typing-extensions>=4.8.0
